requirements.txt
aiogram==3.13.1
aiomysql==0.2.0
python-dotenv==1.1.0
sqlalchemy==2.0.38
PyMySQL==1.1.1